% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{inpop2}
\alias{inpop2}
\title{Predicted Norweigen male population 1998-2022}
\format{A data frame with 18 rows and 5 variables:}
\usage{
inpop2
}
\description{
A dataset containing predicted population numbers for Norweigen men.
\describe{
  \item{rows}{Each row identifies a five year age interval, from 0-4 to 85+}
  \item{columns}{Each column identifies a five year period, from 1998-2002 to 2018-2022}
}
}
\seealso{
Other nordpred_example_data: \code{\link{indata}};
  \code{\link{inpop1}}
}
\keyword{datasets}

