% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.nordpred.R
\name{plot.nordpred}
\alias{plot.nordpred}
\title{plots the predicted rates from a nordpred object}
\usage{
\method{plot}{nordpred}(x, incidence = TRUE, standpop = NULL,
  agegroups = "all", startplot = 1, xlab = "", ylab = "", main = "",
  labels = NULL, ylim = NULL, lty = c(1, 3), col = c(1, 1),
  new = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{nordpred} (see \code{\link{nordpred.object}}}

\item{incidence}{Indicates whether to plot incidence or number of cases}

\item{standpop}{A vector of weights for age standardisation.
Default is no standardisation (crude rates), but using a standardisation
(for the suitable no of age groups) is recommended}

\item{agegroups}{Which agegroups to include}

\item{startplot}{Numeric}

\item{xlab,ylab,main,ylim,lty,col}{Arguments passed to \code{\link{plot}} and \code{\link{par}}}

\item{labels}{character}

\item{new}{Should a new plot be created (\code{new = TRUE} by default) or should new graphs be added to an existing plot (\code{new = FALSE})?}

\item{...}{Arguments passed to \code{\link{plot}}}
}
\value{
object of class \code{nordpred} (see \code{\link{nordpred.object}}).
}
\description{
\code{plot.nordpred} uses nordpred object to plot observed and predicted rates
}
\details{
This function is a method for the generic function plot for class nordpred.
It can be invoked by calling \code{\link{plot}} for an object of the appropriate class,
or directly by calling \code{plot.nordpred} regardless of the class of the object.
For more available options, see \code{\link{plot}}.
For details of the choice of prediction base, significance test for using
recent slope, and for the power5 model, see Moller B., Fekjaer H. et al. (2002),
see references.
}
\section{Note for S-plus}{

Powerlink is made via a special modification in S-PLUS. This works fine
for the point estimates, but the variance estimates found via the glm-objects
are wrong. For variance estimates, we would rather recommend using R.
}
\examples{
# data (Colon cancer for Norwegian males)
indata

# Create dataset with observed and predicted population
inpop <- cbind(inpop1, inpop2)

# Fit model & predict new incidence:
res <- nordpred(indata, inpop, startestage = 5, startuseage = 6,
         cuttrend = c(0, .25, .5, .75, .75))
res2 <- nordpred(indata, inpop, startestage = 5, startuseage = 6,
     cuttrend = c(0, .25, .5, .75, .75), linkfunc = 'poisson')

# Get results with stanardiziotion:
wstand <- c(0.12, 0.1, 0.09, 0.09, 0.08, 0.08, 0.06, 0.06, 0.06, 0.06,0.05,
            0.04, 0.04, 0.03, 0.02, 0.01, 0.005, 0.005)
round(nordpred.getpred(res, incidence = TRUE, standpop = NULL), 2)
round(nordpred.getpred(res, incidence = TRUE, standpop = wstand), 2)

# Plot results:
plot(res, standpop = wstand)

# Plot results with power5 and poisson links:
plot(res2, standpop = wstand)
plot(res, new = FALSE, lty = c(1, 2), standpop = wstand)

# Different cut trend scenarios, using average drift (recent = FALSE):
est <- nordpred.estimate(cases = indata, pyr = inpop, noperiod = 4, startestage = 5)
plot(nordpred.prediction(est, startuseage = 6, cuttrend = c(0,0,0,0,0),
     recent = FALSE), standpop = wstand, new = TRUE)
plot(nordpred.prediction(est, startuseage = 6, cuttrend = c(1,1,1,1,1),
     recent = FALSE), standpop = wstand, new = FALSE, lty = c(1,2))
plot(nordpred.prediction(est,startuseage=6,cuttrend=c(0,.25,.5,.75,.75),
     recent = FALSE), standpop = wstand, new = FALSE, lty = c(1,4))
}
\author{
Harald Fekjaer and Bjorn Moller (Cancer Registry of Norway)
}
\references{
\itemize{
\item A website for nordpred is available at:
\url{http://www.kreftregisteret.no/software/nordpred/}
\item Background for the methods can be found in: Moller B., Fekjaer H., Hakulinen T.,
Sigvaldason H, Storm H. H., Talback M. and Haldorsen T 'Prediction of cancer
incidence in the Nordic countries: Empirical comparison of different approaches'
Statistics in Medicine 2003; 22:2751-2766
\item An application of the function, using all the default settings, can be
found in: Moller B, Fekjaer H, Hakulinen T, Tryggvadottir L, Storm HH, Talback M,
Haldorsen T. Prediction of cancer incidence in the Nordic countries up to the
year 2020. Eur J Cancer Prev Suppl 2002; 11: S1-S96
}
}
\seealso{
Other nordpred: \code{\link{is.nordpred.estimate}},
  \code{\link{nordpred.estimate}};
  \code{\link{is.nordpred}}, \code{\link{nordpred}};
  \code{\link{nordpred.estimate.object}};
  \code{\link{nordpred.getpred}};
  \code{\link{nordpred.object}};
  \code{\link{nordpred.prediction}};
  \code{\link{print.nordpred.estimate}};
  \code{\link{print.nordpred}};
  \code{\link{summary.nordpred}}
}

