% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.nordpred.R
\name{print.nordpred}
\alias{print.nordpred}
\title{Prints a nordpred object}
\usage{
\method{print}{nordpred}(x, digits = 1, ...)
}
\arguments{
\item{x}{An object of class \code{nordpred} (see \code{\link{nordpred.object}})}

\item{digits}{Specifies the number of digits in the tabulation}

\item{...}{ignored}
}
\value{
object of class \code{nordpred} (see \code{\link{nordpred.object}}).
}
\description{
\code{print.nordpred} prints the observed and predicted number of cases in
a nordpred object
}
\section{Note for S-plus}{

Powerlink is made via a special modification in S-PLUS. This works fine
for the point estimates, but the variance estimates found via the glm-objects
are wrong. For variance estimates, we would rather recommend using R.
}
\examples{
# data (Colon cancer for Norwegian males)
indata

# Create dataset with observed and predicted population
inpop <- cbind(inpop1, inpop2)

# Fit model & predict new incidence:
res <- nordpred(indata, inpop, startestage = 5, startuseage = 6,
     cuttrend = c(0, .25, .5, .75, .75))
res2 <- nordpred(indata, inpop, startestage = 5, startuseage = 6,
     cuttrend = c(0, .25, .5, .75, .75), linkfunc = 'poisson')

# Print / get results:
print(res)
nordpred.getpred(res)
summary(res, printpred = FALSE)
}
\author{
Harald Fekjaer and Bjorn Moller (Cancer Registry of Norway)
}
\references{
\itemize{
\item A website for nordpred is available at:
\url{http://www.kreftregisteret.no/software/nordpred/}
\item Background for the methods can be found in: Moller B., Fekjaer H., Hakulinen T.,
Sigvaldason H, Storm H. H., Talback M. and Haldorsen T 'Prediction of cancer
incidence in the Nordic countries: Empirical comparison of different approaches'
Statistics in Medicine 2003; 22:2751-2766
\item An application of the function, using all the default settings, can be
found in: Moller B, Fekjaer H, Hakulinen T, Tryggvadottir L, Storm HH, Talback M,
Haldorsen T. Prediction of cancer incidence in the Nordic countries up to the
year 2020. Eur J Cancer Prev Suppl 2002; 11: S1-S96
}
}
\seealso{
Other nordpred: \code{\link{is.nordpred.estimate}},
  \code{\link{nordpred.estimate}};
  \code{\link{is.nordpred}}, \code{\link{nordpred}};
  \code{\link{nordpred.estimate.object}};
  \code{\link{nordpred.getpred}};
  \code{\link{nordpred.object}};
  \code{\link{nordpred.prediction}};
  \code{\link{plot.nordpred}};
  \code{\link{print.nordpred.estimate}};
  \code{\link{summary.nordpred}}
}

