% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nordpred.prediction}
\alias{nordpred.prediction}
\title{Calculates predictions based on a nordpred.estimate object}
\usage{
nordpred.prediction(nordpred.estimate.object, startuseage, recent,
  cuttrend = c(0, 0.25, 0.5, 0.75, 0.75))
}
\arguments{
\item{nordpred.estimate.object}{A \code{\link{glm}}-object based on
\code{\link{'nordpred.estimate}}.}

\item{startuseage}{Youngest age group which uses regression model as basis
for predicted rates}

\item{recent}{Project average trend or use the slope for the last 10 years?
(If \code{recent = FALSE}, average trend for the whole observation period is
used, if \code{recent = TRUE}, the slope from the last 10 years is used)}

\item{cuttrend}{Cut trend in predictions? (a vector of proportionsof drift
to be cut in each projection period)}
}
\value{
an object of class \code{\link{nordpred.prediction}}.
}
\description{
\code{nordpred.prediction} uses a \code{\link{nordpred.estimate.object}}
 to calculate prediction of cancer incidence and mortality
}
\section{Note for S-plus}{

Powerlink is made via a special modification in S-PLUS. This works fine
for the point estimates, but the variance estimates found via the glm-objects
are wrong. For variance estimates, we would rather recommend using R.
}
\examples{
# Reading data (Colon cancer for Norwegian males)
indata <- read.table("data//colon-men-Norway.txt",header =T,sep=",",row.names=1)
inpop1 <- read.table("data//men-Norway.txt",header =T,sep=",",row.names=1)
inpop2 <- read.table("data//men-Norway-pred.txt",header =T,sep=",",row.names=1)

# Include possible population predictions
inpop <- cbind(inpop1,inpop2)

# Fit model using powerlink (default):
est <- nordpred.estimate(cases=indata,pyr=inpop,noperiod=4,startestage=5)

# Fit model using poisson link:
est2 <- nordpred.estimate(indata,inpop,4,5,linkfunc="poisson")

# Use estimat object to make predictions:
res <- nordpred.prediction(est,startuseage=6,cuttrend=c(0,.25,.5,.75,.75),rece nt=T)
res2 <- nordpred.prediction(est2,startuseage=6,cuttrend=c(0,.25,.5,.75,.75),rec ent=T)

# Get results:
print.nordpred(res)
nordpred.getpred(res)
summary(res,printpred=F)
}
\author{
Harald Fekjaer and Bjorn Moller (Cancer Registry of Norway)
}
\references{
\itemize{
\item A website for nordpred is available at:
\url{http://www.kreftregisteret.no/software/nordpred/}
\item Background for the methods can be found in: Moller B., Fekjaer H., Hakulinen T.,
Sigvaldason H, Storm H. H., Talback M. and Haldorsen T "Prediction of cancer
incidence in the Nordic countries: Empirical comparison of different approaches"
Statistics in Medicine 2003; 22:2751-2766
\item An application of the function, using all the default settings, can be
found in: Moller B, Fekjaer H, Hakulinen T, Tryggvadottir L, Storm HH, Talback M,
Haldorsen T. Prediction of cancer incidence in the Nordic countries up to the
year 2020. Eur J Cancer Prev Suppl 2002; 11: S1-S96
}
}
\seealso{
Other nordpred: \code{\link{nordpred.estimate.object}};
  \code{\link{nordpred.estimate}};
  \code{\link{nordpred.getpred}};
  \code{\link{nordpred.object}}; \code{\link{nordpred}};
  \code{\link{plot.nordpred}};
  \code{\link{print.nordpred.estimate}};
  \code{\link{print.nordpred}};
  \code{\link{summary.nordpred}}
}

