% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.nordpred.R
\name{summary.nordpred}
\alias{summary.nordpred}
\title{Makes a summary of a nordpred object}
\usage{
\method{summary}{nordpred}(object, printpred = TRUE, printcall = FALSE,
  digits = 1, ...)
}
\arguments{
\item{object}{An object of class \code{nordpred} (see \code{\link{nordpred.object}})}

\item{printpred}{Indicates whether to print the observed and predicted number of cases}

\item{printcall}{Indicates whether to print the function call}

\item{digits}{Specifies the number of digits in the tabulation}

\item{...}{ignored}
}
\value{
object of class \code{nordpred} (see \code{\link{nordpred.object}}).
}
\description{
\code{summary.nordpred} uses a \code{nordpred} object (see \code{\link{nordpred.object}})
to summarize
the information
}
\section{Note for S-plus}{

Powerlink is made via a special modification in S-PLUS. This works fine
for the point estimates, but the variance estimates found via the glm-objects
are wrong. For variance estimates, we would rather recommend using R.
}
\examples{
# data (Colon cancer for Norwegian males)
indata

# Create dataset with observed and predicted population
inpop <- cbind(inpop1, inpop2)

# Fit model & predict new incidence:
res <- nordpred(indata, inpop, startestage = 5, startuseage = 6, cuttrend = c(0,.25,.5,.75,.75))
res2 <- nordpred(indata, inpop, startestage = 5, startuseage = 6,
     cuttrend = c(0, .25, .5, .75, .75), linkfunc = 'poisson')

# Print / get results:
print(res)
nordpred.getpred(res)
summary(res, printpred = FALSE)
}
\author{
Harald Fekjaer and Bjorn Moller (Cancer Registry of Norway)
}
\references{
\itemize{
\item A website for nordpred is available at:
\url{http://www.kreftregisteret.no/software/nordpred/}
\item Background for the methods can be found in: Moller B., Fekjaer H., Hakulinen T.,
Sigvaldason H, Storm H. H., Talback M. and Haldorsen T 'Prediction of cancer
incidence in the Nordic countries: Empirical comparison of different approaches'
Statistics in Medicine 2003; 22:2751-2766
\item An application of the function, using all the default settings, can be
found in: Moller B, Fekjaer H, Hakulinen T, Tryggvadottir L, Storm HH, Talback M,
Haldorsen T. Prediction of cancer incidence in the Nordic countries up to the
year 2020. Eur J Cancer Prev Suppl 2002; 11: S1-S96
}
}
\seealso{
Other nordpred: \code{\link{is.nordpred.estimate}},
  \code{\link{nordpred.estimate}};
  \code{\link{is.nordpred}}, \code{\link{nordpred}};
  \code{\link{nordpred.estimate.object}};
  \code{\link{nordpred.getpred}};
  \code{\link{nordpred.object}};
  \code{\link{nordpred.prediction}};
  \code{\link{plot.nordpred}};
  \code{\link{print.nordpred.estimate}};
  \code{\link{print.nordpred}}
}

